Add checker interface
This will be useful to refactor the checkers loop.
